adminclient/requirements.txt
redis
urllib3
pyota[ccurl,pow]
flask
environs
pycrypto
argparse